% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{cna}
\alias{cna}
\title{cna() check count of NAs in a data frame by columh}
\usage{
cna(x)
}
\arguments{
\item{x}{input object to be converted to a dataframe}

\item{factors}{Defaults to 0: (0 indicates no columns are converted to factors)}
}
\description{
cna() check count of NAs in a data frame by columh
}
\examples{
cna(mytable)
x<-cna(mydataframe)
}
\keyword{dataframe}
