% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defl.R
\name{defl}
\alias{defl}
\title{defl() load my default libraries}
\usage{
defl()
}
\description{
defl() load my default libraries
}
\examples{
defl()
}
\keyword{data.table,}
\keyword{default}
\keyword{dplyr,}
\keyword{ggplot2,}
\keyword{library:}
\keyword{magrittr,}
\keyword{readr}
\keyword{reshape2,}
\keyword{tidyr,}
