% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpub.R
\name{ggpub}
\alias{ggpub}
\title{Publication Theme for ggplot}
\usage{
ggpub(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{default size of plot}

\item{base_family}{default font family}
}
\value{
a publication ready ggplot theme
}
\description{
Publication Theme for ggplot
}
