% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf.R
\name{mdf}
\alias{mdf}
\title{mdf() convert an object into a data frame}
\usage{
mdf(x, factors = 0)
}
\arguments{
\item{x}{input object to be converted to a dataframe}

\item{factors}{Defaults to 0: (0 indicates no columns are converted to factors)}
}
\description{
mdf() convert an object into a data frame
}
\examples{
x<-matrix(c(1:12),ncol=3)
x<-mdf(x)
}
\keyword{dataframe}
