% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt.R
\name{mdt}
\alias{mdt}
\title{mdf() convert an object into a data frame}
\usage{
mdt(x, factors = 0)
}
\arguments{
\item{x}{input object to be converted to a data table}

\item{none}{}
}
\description{
mdf() convert an object into a data frame
}
\examples{
x<-matrix(c(1:12),ncol=3)
names(x)<-c('A','B','C')
x<-mdt(x)
}
\keyword{data.table}
