% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcb.R
\name{rcb}
\alias{rcb}
\title{rcb() read excel input from clipboard}
\usage{
rcb(head = TRUE, ...)
}
\arguments{
\item{head}{defaults to TRUE}

\item{...}{additional parameters if used}
}
\description{
rcb() read excel input from clipboard
}
\keyword{clipboard}
\keyword{excel,}
