% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl.R
\name{spl}
\alias{spl}
\title{spl() set plot layout}
\usage{
spl(r = 1, c = 1)
}
\arguments{
\item{...}{additional parameters if used}
}
\description{
spl() set plot layout
}
\keyword{layout,}
\keyword{par}
\keyword{plot,}
