% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcb.R
\name{wcb}
\alias{wcb}
\title{wcb() read excel input from clipboard}
\usage{
wcb(mydata)
}
\arguments{
\item{mydata}{}
}
\description{
wcb() read excel input from clipboard
}
\examples{
somedata<-matrix(c(1:20),ncol=4)
names(somedata)<-c('first','second','third','fourth')
wcb(somedata)
#column names included in the clipboard output
}
\keyword{clipboard}
\keyword{excel,}
